+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support for KMS Software Keys added
 - Open mode and Permission level added to autonomous database resource
 - Support for Flexible Infrastructure - Flexible Memory
+- Support for resizing of boot_volume in `oci_core_instance` resource
 
 ### Fixed
 - Fix logging to provide better feedback to the user to contact service when there are 401 and 500 status codes as response